added linear interpolation as downward extrapolation method.
In cases where ICAR levels existed below the lowest vertical forcing level this code initially used the quantity's value from the lowest forcing level. Now it's extrapolated linerly based on the value in the lowest two forcing levels. This is essentially a copy of the code in vLUT_forcing where the order of the dimensions has been adjusted. Todo: Standard behaviour needs to be preserved, so this should be made optional in the future