Merged origin/feature/Nfromforcing - Added the option N_from_forcing to the linear theory namelist.
If set to true ICAR calculates the atmospheric stability within the ICAR domain from the current state of the forcing data set instead of the ICAR domain. This is in agreement with linear theory where atmospheric stability is a property of the atmospheric background state, which, in case of ICAR, is given by the forcing.