upwind advection - added optional boundary conditions at model top.
Currently the boundary condition employed at the model top is a zero gradient BC placed on all advectable quantities. This commit adds the options to use (1) constant gradient (2) zero value or (3) a zero value BC for the flux divergence. If BC (1) or (2) are chosen the BC is placed on the advectable quantity, while (3) for it is applied to the flux divergence of the topmost vertical layer.

If the new bctop option is set, the boundary condition affects only the microphysics species, if bcthtop is used only the potential temperature is subjected to the altered boundary condition. In the latter case only (1) is allowed as an option.