N can now be read in from a field in the forcing data
The Brunt-Vaisala frequency N can now be read from a field in the
forcingt dataset. To do so N_from_forcing must be set to true and the
name of the field that contains the squared Brunt Vaisala frequency
must be defined in the ICAR options with the option nvar.